getEventsTestsを修正
