+++ springboot/src/test/java/com/asakatu/EventControllerTests.java
@@ -1,8 +1,13 @@
 package com.asakatu;
 
+import com.asakatu.entity.Event;
+import com.asakatu.response.GetEventsListResponse;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.asakatu.entity.UserStatus;
 import com.asakatu.repository.EventRepository;
+import org.hamcrest.core.Is;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -15,9 +20,11 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -51,14 +58,18 @@ public class EventControllerTests extends AbstractTest{
     @WithMockUser
     public void getEventsList() throws Exception {
 		String uri = "/events";
-		MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
-				.accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+		MvcResult mvcResult = mvc.perform(get(uri)
+				.accept(MediaType.APPLICATION_JSON_VALUE)).andDo(print()).andReturn();
 
 		int status = mvcResult.getResponse().getStatus();
 		assertEquals(200, status);
-//		String content = mvcResult.getResponse().getContentAsString();
-//		Event[] eventList = super.mapFromJson(content, Event[].class);
-//		assertTrue(eventList.length > 0);
+		String content = mvcResult.getResponse().getContentAsString();
+
+		TOkResponse<GetEventsListResponse> contentObj = mapper.readValue(content, new TypeReference<TOkResponse<GetEventsListResponse>>(){});
+		List<Event> eventsList = contentObj.getData().getEventsList();
+		Assert.assertThat(eventsList.size(), Is.is(5));
+
 	}
 
 	private UserStatus joinUserStatus() {
@@ -88,4 +99,13 @@ public class EventControllerTests extends AbstractTest{
                 .andExpect(jsonPath("$.data.comment").value("comment"))
                 .andExpect(jsonPath("$.data.message").value("created"));
     }
+}
+
+
+class TOkResponse<T>{
+	private T data;
+
+	public T getData() {
+		return data;
+	}
 }
